+++ packages/app/src/config.ts
@@ -23,5 +23,3 @@ export const IS_DEVELOPMENT = process.env.NODE_ENV !== 'production';
 export const IS_TEST = process.env.NODE_ENV === 'test';
 export const IS_CRX_POPUP =
   IS_CRX && globalThis.location.pathname === CRXPages.popup;
-export const BLOCK_EXPLORER_URL =
-  'https://fuellabs.github.io/block-explorer-v2';
