+++ packages/app/src/systems/Transaction/utils/index.tsx
@@ -7,8 +7,6 @@ import type {
   TxResponse,
 } from '../types';
 
-import { BLOCK_EXPLORER_URL } from '~/config';
-
 export function parseTransaction<T extends TxRequest | TxResponse>(tx: T) {
   return Object.entries(tx).reduce((obj, [key, value]) => {
     const val =
@@ -27,18 +25,6 @@ export function getCoinOutputsFromTx(tx?: TxRequest) {
   return (tx?.outputs ?? []).filter(
     (i) => i.type === OutputType.Coin
   ) as TxOutputCoin[];
-}
-
-export function getBlockExplorerLink({
-  path,
-  providerUrl,
-}: {
-  path: string;
-  providerUrl: string;
-}) {
-  return `${BLOCK_EXPLORER_URL}${path}?providerUrl=${encodeURIComponent(
-    providerUrl
-  )}`;
 }
 
 export * from './error';
