+++ packages/sdk/src/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './blockExplorer';
 export * from './createReadOnly';
 export * from './createUUID';
 export * from './injectFuelWeb3';
